--- a/packages/jest-transform/src/shouldInstrument.ts
+++ b/packages/jest-transform/src/shouldInstrument.ts
@@ -8,13 +8,27 @@
 import * as path from 'path';
 import type {Config} from '@jest/types';
 import {escapePathForRegex} from 'jest-regex-util';
-import {replacePathSepForGlob} from 'jest-util';
+import {globsToMatcher, replacePathSepForGlob} from 'jest-util';
 import micromatch = require('micromatch');
 import type {ShouldInstrumentOptions} from './types';
 
 const MOCKS_PATTERN = new RegExp(
   escapePathForRegex(path.sep + '__mocks__' + path.sep),
 );
+
+const cachedRegexes = new Map<string, RegExp>();
+const getRegex = (regexStr: string) => {
+  if (!cachedRegexes.has(regexStr)) {
+    cachedRegexes.set(regexStr, new RegExp(regexStr));
+  }
+
+  const regex = cachedRegexes.get(regexStr)!;
+
+  // prevent stateful regexes from breaking, just in case
+  regex.lastIndex = 0;
+
+  return regex;
+};
 
 export default function shouldInstrument(
   filename: Config.Path,
@@ -33,15 +47,15 @@ export default function shouldInstrument(
   }
 
   if (
-    !config.testPathIgnorePatterns.some(pattern => !!filename.match(pattern))
+    !config.testPathIgnorePatterns.some(pattern =>
+      getRegex(pattern).test(filename),
+    )
   ) {
     if (config.testRegex.some(regex => new RegExp(regex).test(filename))) {
       return false;
     }
 
-    if (
-      micromatch([replacePathSepForGlob(filename)], config.testMatch).length
-    ) {
+    if (globsToMatcher(config.testMatch)(replacePathSepForGlob(filename))) {
       return false;
     }
   }
@@ -59,10 +73,9 @@ export default function shouldInstrument(
     // still cover if `only` is specified
     !options.collectCoverageOnlyFrom &&
     options.collectCoverageFrom.length &&
-    micromatch(
-      [replacePathSepForGlob(path.relative(config.rootDir, filename))],
-      options.collectCoverageFrom,
-    ).length === 0
+    !globsToMatcher(options.collectCoverageFrom)(
+      replacePathSepForGlob(path.relative(config.rootDir, filename)),
+    )
   ) {
     return false;
   }