I was profiling some Jest runs at Airbnb and noticed that on my
MacBook Pro, we can spend over 2 seconds at Jest startup time in
SearchSource getTestPaths. I believe that this will grow as the size
of the codebase increases.

Looking at the call stacks, it appears to be calling micromatch
repeatedly, which calls picomatch, which builds a regex out of the
globs. It seems that the parsing and regex building also triggers the
garbage collector frequently.

Upon testing, I noticed that the globs don't actually change between
these calls, so we can save a bunch of work by making a micromatch
matcher and reusing that function for all of the paths.

micromatch has some logic internally to handle lists of globs that
may include negated globs. A naive approach of just checking if it
matched any of the globs won't capture that, so I copied and
simplified the logic from within micromatch.

https://github.com/micromatch/micromatch/blob/fe4858b0/index.js#L27-L77

In my profiling of this change locally, this brings down the time of
startRun from about 2000ms to about 200ms.

Before:
![image](https://user-images.githubusercontent.com/195534/83883496-26757e80-a709-11ea-8521-2a71d27ca709.png)

Before, zoomed in:
![image](https://user-images.githubusercontent.com/195534/83971126-26a78280-a89f-11ea-889e-80e925878e43.png)


After:

![image](https://user-images.githubusercontent.com/195534/83883541-37be8b00-a709-11ea-9ee6-f6fdccced6fd.png)

---

**Avoid recreating RegExp instances in regexToMatcher**

After optimizing globsToMatcher, I noticed that there was still a
lot of unnecessary overhead at Jest startup time spent recreating
the same RegExp instances repeatedly. Thankfully, we can be a little
smarter about this and create them all ahead of time and just reuse
them.

On top of my globsToMatcher optimization, this brings the speed of
the ArrayMap in startRun down from about 160ms to about 7ms.

Before:
![image](https://user-images.githubusercontent.com/195534/84039093-5ff6f580-a966-11ea-8eab-282994c45dd5.png)

After:
![image](https://user-images.githubusercontent.com/195534/84039144-72712f00-a966-11ea-9f1f-b4d5e6a97acd.png)

---

**Optimize micromatch and RegExps in shouldInstrument**

I've been profiling running Jest with code coverage at Airbnb, and
noticed that shouldInstrument is called often and is fairly
expensive. It also seems to call micromatch and `new RegExp`
repeatedly, both of which can be optimized by caching the work to
convert globs and strings into matchers and regexes.

I profiled this change by running a set of 27 fairly simple tests.
Before this change, about 6-7 seconds was spent in shouldInstrument.
After this change, only 400-500 ms is spent there. I would expect
this delta to increase along with the number of tests and size of
their dependency graphs.

A typical shouldInstrument call before this change:
![image](https://user-images.githubusercontent.com/195534/84052508-48276d80-a976-11ea-8d6b-a8930d5a9e45.png)

A typical shouldInstrument call after this change (most are actually too fast to even show up in the profiler now):
![image](https://user-images.githubusercontent.com/195534/84052571-60978800-a976-11ea-800d-caf7d58d76fe.png)

---

I was profiling some Jest runs at Airbnb and noticed that on my
MacBook Pro, we can spend over 30 seconds after running Jest with code
coverage as the coverage reporter adds all of the untested files. I
believe that this will grow as the size of the codebase increases.

Looking at the call stacks, it appears to be calling micromatch
repeatedly, which calls picomatch, which builds a regex out of the
globs. It seems that the parsing and regex building also triggers the
garbage collector frequently.

Since this is in a tight loop and the globs won't change between
checks, we can greatly improve the performance here by using
micromatch.matcher.

This optimization reduces the block of time here from about 30s to
about 10s. The aggregated total time of coverage reporter's
onRunComplete goes from 23s to 600ms.

Before:

![image](https://user-images.githubusercontent.com/195534/83884747-fc24c080-a70a-11ea-8440-2c658b3dcce4.png)

Before, zoomed in:
![image](https://user-images.githubusercontent.com/195534/84053653-d819e700-a977-11ea-9597-4b650f4c26cb.png)



After:

![image](https://user-images.githubusercontent.com/195534/83884781-08108280-a70b-11ea-9d63-5329fe726d31.png)

After, zoomed in:
![image](https://user-images.githubusercontent.com/195534/84053716-f08a0180-a977-11ea-8dee-167ddef18ff0.png)


<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

<!-- Please remember to update CHANGELOG.md in the root of the project if you have not done so. -->

## Summary

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->
Motivation: Improve slow Jest startup and runtime speed, particularly when running with coverage

## Test plan

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI. -->

I ran jest in the Airbnb frontend monorepo with and without coverage options, with a path argument.