@SimenB 2020-06-05
@jonschlinkert hiya, thoughts on this?