@lencioni 2020-06-08
To avoid having to stack a bunch of separate PRs, I've added another commit to this one to optimize the startup even more.