@SimenB 2020-06-08
This LGTM, I just want some more eyeballs on it before merging