@AdaskoTheBeAsT 2020-06-15
@cpojer @thymikee @jeysal can you please review this pull request? :) whole community keep fingers crossed for increased performance :)