@7rulnik 2020-06-22
Tested this branch on my company project: duration decreased from 11 minutes to 4 minutes. Seems that it is one of the main reasons for the https://github.com/facebook/jest/issues/9457#issuecomment-646917582 regression.  @lencioni awesome job, thanks!