@SimenB 2020-06-23
Thanks @lencioni! Published as 26.1.0