@lencioni 2020-06-23
In CI we currently spread our tests across 10 separate machines.

When running all tests, the 10 machines on 26.0.1 had runtimes of:

- 10m 28s
- 9m 47s
- 10m 11s
- 9m 59s
- 10m 9s
- 11m 8s
- 11m 26s
- 9m 53s
- 10m 33s
- 13m 39s

On 26.1.0:

- 8m 8s
- 10m 0s
- 9m 28s
- 9m 36s
- 9m 8s
- 9m 50s
- 10m 41s
- 9m 21s
- 10m 8s
- 9m 54s

n=1

Not as much as I had wished for, but this is definitely a nice win for us! Thank you for the reviews and the release!